requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pymysql
passlib
PyJWT
python-multipart
email-validator
